requirements.txt
Flask
google-generativeai
google-generativeai>=0.3.2
